Fixed crash by making this async

+++ runtime/commands/tools/echo.js
@@ -1,9 +1,14 @@
 "use strict";
 
-function run (options)
+async function run (options)
 {
 	BotActions.send(options, options.full);
-	options.message.delete().catch();
+
+	let myMember = options.guild.members.cache.get(VarClient.user.id);
+	let hasPerm = myMember.permissions.has("MANAGE_MESSAGES");
+	if (hasPerm)
+		options.message.delete().catch();
+	// todo: BotTools delete message function
 }
 
 module.exports =
